third_party/blink/web_tests/wpt_internal/route/route-match-pseudo-004.html

<!DOCTYPE html>
<title>:link-to(url-pattern()) serialization</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://issues.chromium.org/issues/436805487">
<style id="style">
  :link-to(  url-pattern(  "*route/other.html"  )  )  {   }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(()=> {
    assert_equals(style.sheet.rules.length, 1);
    assert_equals(style.sheet.rules[0].cssText,
                  ':link-to(url-pattern("*route/other.html")) { }');
  }, "Serialize");
</script>
